Refactor: 폴더 정리 및 Chart Section 디자인 수정
