+++ src/components/Image/Image.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/mixin_var.scss";
+@import "/src/styles/mixin_var.scss";
 
 .rectangleImage {
   position: relative;
@@ -23,5 +23,29 @@
     height: 100%;
     object-fit: cover;
     border-radius: 8px;
+  }
+}
+.roundImage {
+  padding: 5px;
+  border: 1px solid var(--brand-orange);
+  overflow: hidden;
+  border-radius: 50%;
+  &.sm {
+    width: 70px;
+    height: 70px;
+  }
+  &.md {
+    width: 100px;
+    height: 100px;
+  }
+  &.lg {
+    width: 128px;
+    height: 128px;
+  }
+  img {
+    width: 100%;
+    height: 100%;
+    border-radius: 50%;
+    object-fit: cover;
   }
 }
