--- src/components/Image/RoundImage/RoundImage.module.scss
@@ -1,26 +1,0 @@
-@import "../../../styles/mixin_var.scss";
-
-.roundImage {
-  padding: 5px;
-  border: 1px solid var(--brand-orange);
-  overflow: hidden;
-  border-radius: 50%;
-  &.sm {
-    width: 70px;
-    height: 70px;
-  }
-  &.md {
-    width: 100px;
-    height: 100px;
-  }
-  &.lg {
-    width: 128px;
-    height: 128px;
-  }
-  img {
-    width: 100%;
-    height: 100%;
-    border-radius: 50%;
-    object-fit: cover;
-  }
-}
