+++ src/components/Tab/Tab.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/mixin_var.scss";
+@import "/src/styles/mixin_var.scss";
 
 .tabContainer {
   display: flex;
